LPS-61133 migrate internal package
