+++ modules/portal-security/portal-security-sso-facebook-connect/src/main/java/com/liferay/portal/security/sso/facebook/connect/internal/module/configuration/definition/FacebookConnectCompanyServiceConfigurationPidMapping.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.facebook.connect.module.configuration.definition;
+package com.liferay.portal.security.sso.facebook.connect.internal.module.configuration.definition;
 
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
+import com.liferay.portal.security.sso.facebook.connect.constants.FacebookConnectConstants;
 import com.liferay.portal.security.sso.facebook.connect.module.configuration.FacebookConnectConfiguration;
 
 import org.osgi.service.component.annotations.Component;
@@ -23,12 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Mika Koivisto
  */
 @Component
-public class FacebookConnectCompanyServiceConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
+public class FacebookConnectCompanyServiceConfigurationPidMapping
+	implements ConfigurationPidMapping {
 
 	@Override
 	public Class<?> getConfigurationBeanClass() {
 		return FacebookConnectConfiguration.class;
+	}
+
+	@Override
+	public String getConfigurationPid() {
+		return FacebookConnectConstants.SERVICE_NAME;
 	}
 
 }
